test: refactor denormalization helpers
