+++ src/core/client/stream/test/fixtures.ts
@@ -1,3 +1,10 @@
+import {
+  denormalizeAsset,
+  denormalizeAssets,
+  denormalizeComment,
+  denormalizeComments,
+} from "talk-framework/testHelpers";
+
 export const users = [
   {
     id: "user-0",
@@ -12,52 +19,6 @@ export const users = [
     username: "Isabelle",
   },
 ];
-
-function denormalizeComment(comment: any, parents: any[] = []) {
-  const replyNodes =
-    (comment.replies &&
-      comment.replies.edges.map((edge: any) =>
-        denormalizeComment(edge, [...parents, comment])
-      )) ||
-    [];
-  const repliesPageInfo = (comment.replies && comment.replies.pageInfo) || {
-    endCursor: null,
-    hasNextPage: false,
-  };
-  return {
-    ...comment,
-    replies: { edges: replyNodes, pageInfo: repliesPageInfo },
-    replyCount: replyNodes.length,
-    parentCount: parents.length,
-    parents: {
-      edges: parents,
-      pageInfo: { startCursor: null, hasPreviousPage: false },
-    },
-  };
-}
-
-function denormalizeComments(commentList: any[]) {
-  return commentList.map(c => denormalizeComment(c));
-}
-
-function denormalizeAsset(asset: any) {
-  const commentNodes =
-    (asset.comments &&
-      asset.comments.edges.map((edge: any) => denormalizeComment(edge, 0))) ||
-    [];
-  const commentsPageInfo = (asset.comments && asset.comments.pageInfo) || {
-    endCursor: null,
-    hasNextPage: false,
-  };
-  return {
-    ...asset,
-    comments: { edges: commentNodes, pageInfo: commentsPageInfo },
-  };
-}
-
-function denormalizeAssets(assetList: any[]) {
-  return assetList.map(a => denormalizeAsset(a));
-}
 
 export const comments = denormalizeComments([
   {
